20230227 update
메인메뉴, 치킨겜 수정

+++ Assets/KDH/Scripts/MainMenu/ScoreManager.cs
@@ -32,7 +32,7 @@ public class ScoreManager : MonoBehaviour
         }
     }
 
-    public void SetGame2Score(float _score)
+    public void SetGame2Score(int _score)
     {
         if (_score > game2ScoreMax)
         {
